refactor(desktop): extract markdown types and utilities
Move markdown renderer types and shared utility helpers into markdown submodules without changing their implementations.
